Refactor: Simplified text ranges
Most parts of the code actually don't need a text range,
just a start and an end location. So, the code now only
gets those.

Because of that, the code also doesn't need most of the
original functionality of ranges anymore, and some of
that could be moved to locations.

Last, removed the bound functions and created abstract
base classes, since this seems to be more suitable
and better design.