Parsing Headings and Sections
The HeadingParser can parse headings, and the section parser
now knows when a heading is part of the current section,
when it should start a new inner section and when it should
end the current section.

When the section parser ends the current section, the calling
parser (another section parser or the ContainerParser) will
start the next section.