Parse simple strong emphasis
Simple strong emphasis (which contains only plain text)
can be parsed now. The parser will only return a strong
emphasis when it is at the start of the parsed range.
It can also find the next possible start of a strong
emphasis.

So, a calling parser can now search for the next possible
start, then try to parse the inline. If it worked, it can
add the text before it as plain text and the inline. If
not, it has to search further for a possible start.

I guess all inlines will have to provide this information
(next possible start)...