Parse emphasis and strong emphasis
Parsing emphasis and strong emphasis works. BUT there
are several known incompatibilites with GfM; Especially,
when it comes to nesting delimited inlines but also when
searching for the shortest span.

Being 100% compatible here would complicat initial parsing
and update parsing. And, when nesting emphasis or when there
are ambiguouties, it is better to use different symbols
(*, _) anyway.