Added parent element
Elements now have a parent element, and MfMElement implements
setting / removing the element. For this, it assumes that any
element in the content array is also an MfMElement.

The parent element will later be used to create a custom context
that can be passed down the tree, which marmota.live will use
to render the elements.